Moved food route into its own module
